+++ src/qinfer/tests/base_test.py
@@ -36,7 +36,7 @@ import abc
 import numpy as np
 from numpy.testing import assert_equal, assert_almost_equal
 import unittest
-from qinfer import Domain
+from qinfer import Domain, FiniteOutcomeModel
 
 ## FUNCTIONS ##################################################################
 
@@ -74,6 +74,41 @@ def test_model(model, prior, expparams, stream=sys.stderr):
     runner.run(test)
 
 ## CLASSES ####################################################################
+
+class MockModel(FiniteOutcomeModel):
+    """
+    Two-outcome model whose likelihood is always 0.5, irrespective of
+    model parameters, outcomes or experiment parameters.
+    """
+
+    def __init__(self):
+        super(MockModel, self).__init__()
+    
+    @property
+    def n_modelparams(self):
+        return 2
+        
+    @staticmethod
+    def are_models_valid(modelparams):
+        return np.ones((modelparams.shape[0], ), dtype=bool)
+        
+    @property
+    def is_n_outcomes_constant(self):
+        return True
+        
+    def n_outcomes(self, expparams):
+        return 2
+
+        
+    @property
+    def expparams_dtype(self):
+        return [('a', float), ('b', int)]
+        
+    
+    def likelihood(self, outcomes, modelparams, expparams):
+        super(MockModel, self).likelihood(outcomes, modelparams, expparams)
+        pr0 = np.ones((modelparams.shape[0], expparams.shape[0])) / 2
+        return FiniteOutcomeModel.pr0_to_likelihood_array(outcomes, pr0)
 
 class DerandomizedTestCase(unittest.TestCase):
